+++ src/dev.rs
@@ -528,7 +528,7 @@ impl<F: FieldExt> MockProver<F> {
         }
 
         if instance.len() != cs.num_instance_columns {
-            return Err(Error::IncompatibleParams);
+            return Err(Error::InvalidInstances);
         }
 
         let instance = instance