+++ src/plonk/error.rs
@@ -5,9 +5,8 @@ pub enum Error {
     /// This is an error that can occur during synthesis of the circuit, for
     /// example, when the witness is not present.
     SynthesisError,
-    /// The structured reference string or the parameters are not compatible
-    /// with the circuit being synthesized.
-    IncompatibleParams,
+    /// The provided instances do not match the circuit parameters.
+    InvalidInstances,
     /// The constraint system is not satisfied.
     ConstraintSystemFailure,
     /// Out of bounds index passed to a backend