+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to Rust's notion of
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- `halo2::plonk::Error` has been overhauled:
+  - `Error::IncompatibleParams` is now `Error::InvalidInstances`.
+
 ### Removed
 - `halo2::arithmetic::BatchInvert` (use `ff::BatchInvert` instead).